Refactor generation scripts for proper reusability
