+++ server/db/generateReviews.js
@@ -1,20 +1,18 @@
-const mongoose = require('mongoose');
+const { review } = require('./schemas.js');
+const { db } = require('./db.js');
 
-mongoose.connect('mongodb://localhost/hippos');
+// Helper functions
+const randIdx = length => Math.floor(Math.random() * length);
 
-const reviewSchema = new mongoose.Schema({
-  restaurantId: Number,
-  name: String,
-  image: String,
-  rank: String,
-  date: String,
-  stars: String,
-  description: String,
-});
+const toInclude = (prob, includeVal, notIncludeVal) => {
+  if (Math.floor(Math.random() * 100) < prob) return includeVal;
+  return notIncludeVal;
+};
 
-const review = mongoose.model('review', reviewSchema);
+// Random generation functions
 
-const db = mongoose.connection;
+
+// Store data functions
 db.on('error', console.error.bind(console, 'connection error:'));
 db.once('open', () => {
   console.log('Database connected! Ready to insert data...');
